+++ Package.swift
@@ -1,4 +1,4 @@
-// swift-tools-version:5.3
+// swift-tools-version:5.9
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
@@ -8,6 +8,8 @@ let package = Package(
     platforms: [
         .macOS(.v11),
         .iOS(.v12),
+        .tvOS(.v12),
+        .visionOS(.v1)
     ],
     products: [
         .library(