+++ Sources/Highlighter/Shims.swift
@@ -12,11 +12,11 @@ import Foundation
 /**
  Basic imports and typealiases.
  */
-#if os(OSX)
+#if os(macOS)
 import AppKit
 public typealias HRColor = NSColor
 public typealias HRFont  = NSFont
-#elseif os(iOS)
+#else
 import UIKit
 public typealias HRColor = UIColor
 public typealias HRFont  = UIFont
@@ -26,18 +26,11 @@ public typealias HRFont  = UIFont
 /**
  Set type aliases according to which Swift is being run and,
  in the second case, if we're running on iOS
-
- NOTE This is probably unnecessary now since the Swift Package
-      mandates Swift 5.3
  */
-#if swift(>=4.2)
 public typealias AttributedStringKey = NSAttributedString.Key
-#else
-public typealias AttributedStringKey = NSAttributedStringKey
-#endif
 
-#if swift(>=4.2) && os(iOS)
-public typealias TextStorageEditActions = NSTextStorage.EditActions
-#else
+#if os(macOS)
 public typealias TextStorageEditActions = NSTextStorageEditActions
+#else
+public typealias TextStorageEditActions = NSTextStorage.EditActions
 #endif