+++ Sources/Highlighter/Theme.swift
@@ -111,7 +111,7 @@ public class Theme {
         self.fontSize = font.pointSize
 
         // Generate the bold and italic variants
-#if os(iOS) || os(tvOS)
+#if os(iOS) || os(tvOS) || os(visionOS)
         let boldDescriptor    = UIFontDescriptor(fontAttributes: [UIFontDescriptor.AttributeName.family:font.familyName,
                                                                   UIFontDescriptor.AttributeName.face:"Bold"])
         let italicDescriptor  = UIFontDescriptor(fontAttributes: [UIFontDescriptor.AttributeName.family:font.familyName,