+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 - 1.2.0 *Unreleased*
     - Add line numbering facility.
+    - Add preliminary VisionOS support.
     - Migrate minimum supported macOS version to Big Sur (11).
+    - Update baseline Swift version to 5.9.
 - 1.1.7 *23 July 2025*
     - Update to the latest version of Highlight.js 11.11.1.
 - 1.1.6 *30 August 2024*